+++ py_basic_commands/__init__.py
@@ -15,4 +15,3 @@ from .basic_commands import chunker
 from .basic_commands import create_json
 from .basic_commands import read_json
 from .basic_commands import write_json
-from .basic_commands import fd